Restructure folders e.g. s/Binary.HttpBridge/HttpBridge.Binary/
- Flatten the the folder-structure to make it easier to comprehend in a file-tree-view of an editor

- [x] I have flipped {Binary,Transaction,Compatibility,Network} and {HttpBridge,Jormungandr} such that we now how e.g. `Cardano.Wallet.HttpBridge.Binary`
- [x] `Cardano.Wallet.Network.HttpBridge.Api` is now `Cardano.Wallet.HttpBridge.Api` (now in the same folder as `Cardano.Wallet.HttpBridge.Network`)
- [ ] I have *not* touched the `lib/{http-bridge,jormungandr}/test/data` structure. It is more complex.
