+++ lib/http-bridge/src/Cardano/Wallet/HttpBridge/Compatibility.hs
@@ -10,7 +10,7 @@
 -- (Byron, Shelley-Rust, Shelley-Haskell) and isolate the bits that vary between
 -- those backends.
 
-module Cardano.Wallet.Compatibility.HttpBridge
+module Cardano.Wallet.HttpBridge.Compatibility
     ( -- * Target
       HttpBridge
     ) where
@@ -19,7 +19,7 @@ import Prelude
 
 import Cardano.Environment.HttpBridge
     ( Network (Mainnet, Staging, Testnet), network, protocolMagic )
-import Cardano.Wallet.Binary.HttpBridge
+import Cardano.Wallet.HttpBridge.Binary
     ( encodeAddress, encodeProtocolMagic, encodeTx )
 import Cardano.Wallet.Primitive.AddressDerivation
     ( KeyToAddress (..), getKey )
