+++ lib/http-bridge/src/Cardano/Wallet/HttpBridge/Transaction.hs
@@ -3,7 +3,7 @@
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE TypeApplications #-}
 
-module Cardano.Wallet.Transaction.HttpBridge
+module Cardano.Wallet.HttpBridge.Transaction
     ( newTransactionLayer
     ) where
 
@@ -11,9 +11,9 @@ import Prelude
 
 import Cardano.Environment.HttpBridge
     ( Network (..), ProtocolMagic (..), network, protocolMagic )
-import Cardano.Wallet.Binary.HttpBridge
+import Cardano.Wallet.HttpBridge.Binary
     ( toByteString )
-import Cardano.Wallet.Compatibility.HttpBridge
+import Cardano.Wallet.HttpBridge.Compatibility
     ( HttpBridge )
 import Cardano.Wallet.Primitive.AddressDerivation
     ( Depth (AddressK), Key, Passphrase (..), XPrv, XPub, getKey, publicKey )
